app_step_functions/requirements.txt
boto3==1.28.26
botocore==1.31.26
jmespath==1.0.1
Pillow==10.0.0
python-dateutil==2.8.2
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.16
